Er selma sur? nej

+++ src/data/make_dataset.py
@@ -14,9 +14,6 @@ from PIL import Image
 from torchvision import transforms
 import torch
 from sklearn.model_selection import train_test_split
-import numpy as np
-import matplotlib.pyplot as plt
-from torchvision.transforms.transforms import Resize
 
 # Kornia
 import kornia as K
@@ -34,7 +31,7 @@ def findFileExtension(filename):
 
     extension = file_split[-1]
     filename = file_split[0]
-
+  
     return filename, extension
 
 def main():
@@ -48,14 +45,14 @@ def main():
 
     # TODO: create method how to get raw data
     # TODO: make check to see if the folder structure is there
-
+    
     # TODO: use kornia to do image augmentation
 
     preprocess = transforms.Compose(
         [
             transforms.Resize((224, 224)),
             #transforms.ToTensor(),
-            #transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
+            #transforms.Normalize((0.5,), (0.5,))
         ]
     )
 
@@ -78,8 +75,6 @@ def main():
             gauss = K.filters.GaussianBlur2d((5, 5), (4, 4))
             blur = gauss(processed.float())
 
-            # Rotation
-
             # Make tensor to numpy before adding to lists
             # Add original image and argumented imges
             data.append(processed.detach().numpy())
@@ -95,16 +90,17 @@ def main():
             cv2.imwrite(os.path.join(path+dir , filename), img_original)
 
             # split the name and extension of the file
-            # TODO make the fucking def findFileExtension work
+            #filename, extension = findFileExtension(str(filename)) 
+
             file_split = filename.split(".")
 
             extension = file_split[-1]
-            filename = file_split[0]
+            filename = file_split[0]     
 
             cv2.imwrite(os.path.join(path+dir , filename+'_blur'+'.'+extension), img_blur)
-
+    
     data    = torch.tensor(data)
-    targets = torch.tensor(targets).float()
+    targets = torch.tensor(targets)
 
     X_train, X_test, y_train, y_test = train_test_split(data, targets, 
                                                         test_size=0.2, 
@@ -133,7 +129,7 @@ def main():
     axs[1].imshow(blur)
 
     plt.show()
-    ###
+    ###  
 
 if __name__ == '__main__':
     log_fmt = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
@@ -150,7 +146,7 @@ if __name__ == '__main__':
     parser.add_argument('--init', action='store_true')
     args = parser.parse_args()
     init_value = args.init
-
+    
     if init_value:
         try:
             if not (path.exists('brain_tumor_dataset/raw/')):
